+++ src/test/java/wooteco/subway/path/PathAcceptanceTest.java
@@ -5,7 +5,7 @@ import static org.springframework.restdocs.restassured3.RestAssuredRestDocumenta
 import static wooteco.subway.auth.AuthAcceptanceTest.로그인되어_있음;
 import static wooteco.subway.auth.AuthAcceptanceTest.회원_등록되어_있음;
 import static wooteco.subway.line.LineAcceptanceTest.지하철_노선_등록되어_있음;
-import static wooteco.subway.section.SectionAcceptanceTest.지하철_구간_등록되어_있음;
+import static wooteco.subway.line.SectionAcceptanceTest.지하철_구간_등록되어_있음;
 import static wooteco.subway.station.StationAcceptanceTest.지하철역_등록되어_있음;
 
 import com.google.common.collect.Lists;
